requirements.txt
aenum==3.1.15
aiohttp==3.9.4
aiomysql==0.2.0
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==5.1.1
Deprecated==1.2.14
et_xmlfile==2.0.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.1
frozenlist==1.4.1
future==1.0.0
google-ai-generativelanguage==0.6.1
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.22.0
Hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
ipython==8.18.1
iso8601==1.1.0
jedi==0.19.1
Jinja2==3.1.3
line-bot-sdk==3.10.2
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
multidict==6.0.5
mysql-connector-python==8.3.0
numpy==1.26.4
openpyxl==3.1.5
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
priority==2.0.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyMySQL==1.1.0
pyparsing==3.1.2
pypika-tortoise==0.1.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-multipart==0.0.9
pytz==2024.1
requests==2.31.0
rsa==4.9
sendgrid==6.11.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starkbank-ecdsa==2.2.0
starlette==0.37.2
taskgroup==0.0.0a4
tomli==2.0.1
tortoise-orm==0.20.0
tqdm==4.66.2
traitlets==5.14.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
